@@ -1,20 +1,22 @@
 package jpabook.jpashop.domain.entity;
 
+import jakarta.persistence.*;
 import jpabook.jpashop.domain.entity.item.Item;
 import lombok.Getter;
 import lombok.Setter;
 
-import javax.persistence.*;
 import java.util.ArrayList;
 import java.util.List;
 
-import static javax.persistence.FetchType.*;
+import static jakarta.persistence.FetchType.LAZY;
+
 
 @Entity
 @Getter @Setter
 public class Category {
 
-    @Id @GeneratedValue
+    @Id
+    @GeneratedValue
     @Column(name = "category_id")
     private Long id;
 
