@@ -1,11 +1,11 @@
 package jpabook.jpashop.domain.entity;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
+import jakarta.validation.constraints.NotEmpty;
 import lombok.Getter;
 import lombok.Setter;
 
-import javax.persistence.*;
-import javax.validation.constraints.NotEmpty;
+import jakarta.persistence.*;
 import java.util.ArrayList;
 import java.util.List;
 
