@@ -1,5 +1,6 @@
 package jpabook.jpashop.api;
 
+import jakarta.validation.Valid;
 import jpabook.jpashop.domain.entity.Member;
 import jpabook.jpashop.service.MemberService;
 import lombok.AllArgsConstructor;
@@ -7,7 +8,6 @@ import lombok.Data;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
-import javax.validation.Valid;
 import java.util.List;
 import java.util.stream.Collectors;
 
