@@ -1,9 +1,9 @@
 package jpabook.jpashop.controller;
 
+import jakarta.validation.constraints.NotEmpty;
 import lombok.Getter;
 import lombok.Setter;
 
-import javax.validation.constraints.NotEmpty;
 
 @Getter @Setter
 public class MemberForm {
