@@ -1,8 +1,8 @@
 package jpabook.jpashop.domain.entity;
 
+import jakarta.persistence.Embeddable;
 import lombok.Getter;
 
-import javax.persistence.Embeddable;
 
 @Embeddable
 @Getter
