@@ -1,11 +1,11 @@
 package jpabook.jpashop.domain.entity.item;
 
+import jakarta.persistence.*;
 import jpabook.jpashop.domain.entity.Category;
 import jpabook.jpashop.exception.NotEnoughStockException;
 import lombok.Getter;
 import lombok.Setter;
 
-import javax.persistence.*;
 import java.util.ArrayList;
 import java.util.List;
 
