@@ -1,10 +1,10 @@
 package jpabook.jpashop.repository;
 
+import jakarta.persistence.EntityManager;
 import jpabook.jpashop.domain.entity.Member;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
-import javax.persistence.EntityManager;
 import java.util.List;
 
 @Repository
