@@ -1,10 +1,10 @@
 package jpabook.jpashop.repository.order.simplequery;
 
+import jakarta.persistence.EntityManager;
 import jpabook.jpashop.domain.dto.OrderSimpleQueryDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
-import javax.persistence.EntityManager;
 import java.util.List;
 
 /**
